requirements.txt
boto3==1.34.126
findspark==2.0.1
pyspark==3.5.1
python-dotenv==1.0.1
snowflake_connector_python==3.10.1
